Linebot/requirements.txt
python-dotenv
fastapi[all]
line-bot-sdk
pyngrok
requests